fix operator parsing in expr2
Due to a bug, we were currently parsing left-/non-associative operators
as if they were right-associative, so:

  let x = 1,2,3
    ==>
  let x = (1,(2,3))

The reason is that the grammar was effectively:

  expr0 := 'let' ... | ... | expr1
  expr1 := expr2 (op expr0)*
  expr2 := (applications, literals, parens, etc.)
  op    := ',' | ';' | ...

Note the use of expr0 in expr1 - this is to allow things like

  f (); let x = 1 in ...
  2 + match z with

The problem is that if the RHS contains more operators than they are
incorrectly consumed by the expr0 rather than being handled "properly"
in the grammar.

As a temporary fix we change the expr0 to be instead expr2 - this fixes
the original issue at the cost of forbidding expressions like this,
which is especially inconvenient in the case of semicolons, etc.