parse (func Constructor)
The issue is that the previous zero-argument constructor case lived in
the same production level as the production for application chains -
expr2.

But since arguments were expected to be an *expr3*, we weren't
correctly parsing something like (func Constructor). My solution is to
add an additional case for nullary constructors in expr3, which results
in code duplication but is probably the correct logic.

In principle this would also be needed for patterns to handle things
like (fun Zero -> ...), but since I don't expect to need algebraic types
with a single nullary constructor (which would make them equivalent to
unit), I haven't included this.