create uvars in 'let' head patterns at (lvl + 1)
Previously I thought these uvars should be created at the current level (not
the incremented level) for some reason. But as it turns out this prevents any
generalization from happening, because we unify the head with (some type
expression involving) the argument and result types, and as a result their
level becomes also forced to the current level.