replace '>>= fun ...' with '>>= (fun ...)'
We need to do this because at the moment we don't support having expr0-level
expressions on the right hand side of an operator - because, since a chain of
operations is an expr0 (and therefore an expr1), this would effectively force
every operator to be right-associative. This can be fixed by introducing some
kind of special case (either a bool parameter to expr0, or extracting the other
variants in expr0 into some kind of combinator).