use consistent type variables across patterns
Previously, upon visiting any pattern we would first find all the type
variables that are bound by the pattern, and create a fresh UVar for each one.
This was more relaxed than the behavior in OCaml, since it would allow things
like:

  let f (x : 'a) (y : 'a) : 'a = int_of_string x + int_of_char y

Here every occurrence of 'a would be considered different, allowing them to
unify with string, char, and int respectively. Here we fix this behavior by
introducing the notion of a "dynamic tyvar scope": a scope that lazily comes up
with fresh UVars as it is asked for them. By reusing the same tvs across
multiple patterns, we can ensure that the uvars are consistent.

Specifically, we currently create tvss in the following situations:

- The same tvs is used across all patterns in a `match` expression.
- The same tvs is used across a whole group of `let` bindings.
- A new tvs is created for the parameters to a `fun`.

It seems that in OCaml, type variables are actually scoped to the expression
as well, so something like

  fun (x : 'a) -> match string_of_int x with
                  | y : 'a -> ()

is an error. I haven't implemented this yet, as it would require putting type
variables in the ctx.