reduce the number of temporaries emitted during expression compilation
Previously

  go_expr : expr -> string

had the following contract:
- It emits (as a side effect) lines that define temporaries needed for
  the expression being compiled, including performing any side effects.
- It returns an expression which can be used arbitrarily many times to
  refer to the result of evaluating the expression.

This results in needing to emit temporaries in many places; for example,
the result of a function call or a constructor would need to be saved to
a temporary to uphold the second part of the contract.

We change this to instead be:

  go_expr : expr -> bool * string

with the following contract:
- It emits (as a side effect) lines that define temporaries needed for
  the expression, including performing *some* of the side effects.
- It returns (b, e), where `e` is an expression that can be used *once*
  to compute the result.
- If `b` is true, then it is OK for additional side effects to occur
  in between the emitted lines and the evaluation of `e`.
  (This is referred to as `e` being "pure".)

This system has the advantage that, when we compile an application
`(e1 e2)`, we don't need temporaries for `e1` and `e2` if at least one
of these expressions is pure. If they are both impure, we can emit a
temporary for just `e1`.

This reduces the size of the compiled code from about 10000 lines to
about 7500 lines at time of writing.