change how the entry point works in Chez
Chez doesn't provide a way to say "load this file and then run this
script" from the command line, only "load this file before entering the
REPL." Previously we fixed this by temporarily launching a Chez REPL and
then immediately exiting it. But this causes the command-line arguments
to not be passed correctly, so instead we change compat_chez.scm to be
responsible for loading both the prelude *and* the compiled code.