elab: support elaborating records
OCaml actually has somewhat weird behavior for how it resolves record
names when it doesn't have a result type available. The goal is to allow
the user to qualify some or all of the field names.

Since we don't support qualified record field names at the moment, we do
something simpler. The first field of the record is looked up using the
context, which is enough to tell us the type constructor that we should
have. Based on this, we can look up the rest of the fields directly.