elab: fix typing rule for lists
In a longstanding bug, the typing rule we implemented for lists was
actually

    e1 : t   ...   eN : t
  -------------------------
    [e1; e2; ...; eN] : t

The type at the bottom should be `t list`. The reason this bug hadn't
been found yet is that, at the moment, we still only support empty list
literals, so `t` would stay as an unsolved meta and get unified to the
correct type later.