elab: remove type annotations from `new_uvars`
The way OCaml handles type variables is really weird IMO. It seems to
have decided that a fresh tyvar scope is not created for this local
function `new_uvars`. Consequently, when we write 'a, it assumes that
this 'a is scoped to the entire function `new_elaborator`, and it
actually serves as a metavariable which gets instantiated to one
concrete type. This prevents us from reusing it in both the pattern and
expression code.