elab: change the order in which we type check bindings
For a binding like

  let p : annot = rhs

Originally we would infer the type of `p`, check the type of `rhs`
against `annot`, and then unify the resulting two types.

Now, we check the type of `rhs` against `annot` and then check `p`
against the result.

This seems like the obvious thing to do, but in the situation where
`annot` is omitted, it means we are unable to propagate information from
the pattern into the expression. This means that unfortunately something
like this is not allowed:

  let ((a : Module.out_of_scope_record), ()) =
      ({ some_field = 1; ... }, ())

If the annotation was outside the pattern, then we would visit it before
the expression, so by the time we encountered the record introduction,
we would already know that the result type was `out_of_scope_record` and
be able to check it properly.

Instead, we reject this binding now. I think this is a reasonable
compromise and helps us be consistent with the user's expectations with
`match`, where the scrutinized expression is also visited before the
pattern.