compile_scheme: don't suffix constructor names with their IDs
Constructor names are compiled to symbols, which comply with OCaml's
syntax for constructors, meaning they start with a capital letter and
therefore can't conflict with anything in Scheme. They also won't
conflict with each other, because we don't allow multiple constructors
with the same name within a single ADT (actually, we do, but we
shouldn't).